+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+
+## [Unreleased] 
+
+### Added
+
+- add dcterms:source and plugin provenance
+
 ## [2.0.1] 2025-02-19
 
 ### Changed